player command-line options
Options for everything that can be configured
in the startup dialog, and then some:
--hide-dialog: don't show startup dialog
--time-logging: enable time logging (old option)
--exit-after-precalc: exactly what it says
--fs / --win: fullscreen / windowed mode
--loop / --no-loop: enable / disable looping
--vsync / --no-vsync: enable / disable VSync
--precache / --no-precache: enable / disable precaching
--no-aa / --2x / --4x / --8x / --16x: set MSAA sampling
--640x480 / --1920x1080 / ...: set resolution
--1280x720x50 / ...: set resolution and refresh rate
--16/9 / --4:3 / ...: set aspect ratio

Most options have a few aliases for convenience;
look into Player/CommandLineOptions.cs for a full list.

Dashes and underscores are ignored, i.e. "--no_aa"
and "--noaa" are synonyms for "--no-aa".

Some mild refactoring in ContextSettings, StartUpDialog
and the Player app itself were necessary for this.
Mainly, ContextSettings are now created only once
in the Player app, and both CommandLineOptions and
StartUpDialog get a chance to modify it.