possibility to customize player loading screen
The look of the player's loading screen can now be customized
by defining a custom operator that draws it and defining a
property called "LoaderProgressOperator" with the operator's
meta ID in ProjectSettings.json. The operator will then be
rendered instead of the normal loading bar, with t=0 seconds
at the start of loading and t=1 second at the end.