changed loader time mapping to -1...0
The time for the custom loader operator is now mapped
differently: instead of 0...1, it's -1...0. This makes
it more natural to work with: Now, the loader is conceptually
running in the second just before the demo (except that this
"second" may in reality take much longer than that). It no
longer overlaps part of the timeline. This also makes seamless
transitions from the loader into the first scene easier.